Extract duplicated code to HasOneBlockchainThroughCurrency (closes #1593)
