+++ app/models/payment_address.rb
@@ -12,13 +12,13 @@ class PaymentAddress < ActiveRecord::Base
   serialize :details, JSON
 
   before_validation do
-    next unless blockchain_api&.supports_cash_addr_format? && address?
-    self.address = CashAddr::Converter.to_cash_address(address)
+    next if blockchain_api&.case_sensitive?
+    self.address = address.try(:downcase)
   end
 
   before_validation do
-    next if blockchain_api&.case_sensitive?
-    self.address = address.try(:downcase)
+    next unless blockchain_api&.supports_cash_addr_format? && address?
+    self.address = CashAddr::Converter.to_cash_address(address)
   end
 
   def enqueue_address_generation
